Move generating the requirements into its own service
